requirements.txt
uvicorn[standard]>=0.29.0
fastapi>=0.111.0
hvac>=2.2.0
python-jose[cryptography]>=3.3.0
pydantic>=2.6.0
pydantic-settings>=2.5.0
fastapi-mcp>=0.4.0
httpx>=0.27.0
pytest>=8.2.0
pytest-cov>=4.1.0
prometheus-client>=0.20.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-exporter-otlp>=1.27.0
